+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 2025-09-09
 - #1646 Implements `eth_estimateGas` RPC method for the EVM module. This allows clients to estimate the gas required for transaction execution before submitting them to the network.
+- #1650 Adds an optional `state_cache_size` param to the `storage` section of the rollup config.toml file. 
 - #1645 **BREAKING CHANGE** The `PerBlockCache` trait has been updated. The `put_cached and get_cached` methods now accept an optional `SlotKey`. For backward compatibility, None should be passed when a SlotKey is not provided.
 
 # 2025-09-08